[Local GC] Scaffolding for loading a standalone GC (#11242)
* Configure the build system to build a CoreCLR capable of loading a standalone GC

* Proto-implementation of dynamic GC loading

* Build the GC with the VM's CMakeLists when doing a non-standalone build of the GC

* [Local GC] Introduce a new feature define, FEATURE_STANDALONE_GC_ONLY, to be used by the CI to explicitly test local GC dynamic loading code paths

* Fix the FEATURE_STANDALONE_GC_ONLY build for unix linkers

* Rebase against master

* Code review feedback: use the existing Unix exports file
