+++ src/CMakeLists.txt
@@ -149,7 +149,9 @@ add_subdirectory(utilcode)
 add_subdirectory(gcinfo)
 add_subdirectory(coreclr)
 add_subdirectory(jit)
-add_subdirectory(gc)
+if(FEATURE_STANDALONE_GC)
+  add_subdirectory(gc)
+endif(FEATURE_STANDALONE_GC)
 add_subdirectory(vm)
 add_subdirectory(md)
 add_subdirectory(debug)
