--- src/gc/dac/CMakeLists.txt
@@ -1,2 +1,0 @@
-include(${CLR_DIR}/dac.cmake)
-add_library_clr(gc_dac STATIC ${GC_SOURCES_DAC})
