--- src/gc/wks/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_library_clr(gc_wks STATIC ${GC_SOURCES_WKS})
